requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests
ddgs
psycopg2-binary==2.9.9
pywebview==5.1

